+++ src/commands/hardis/scratch/push.ts
@@ -44,7 +44,7 @@ export default class SourcePush extends SfdxCommand {
 
   public async run(): Promise<AnyJson> {
     const debugMode = this.flags.debug || false;
-    await forceSourcePush(this.org.getUsername(),this, debugMode);
+    await forceSourcePush(this.org.getUsername(), this, debugMode);
     // Return an object to be displayed with --json
     return { outputString: "Pushed local git branch in scratch org" };
   }